Rename seeds directory to seeders and update namespaces
