+++ app/Console/Stubs/Modules/listeners/install.stub
@@ -41,7 +41,7 @@ class FinishInstallation
     {
         Artisan::call('company:seed', [
             'company' => company_id(),
-            '--class' => 'Modules\$STUDLY_NAME$\Database\Seeds\$STUDLY_NAME$DatabaseSeeder',
+            '--class' => 'Modules\$STUDLY_NAME$\Database\Seeders\$STUDLY_NAME$DatabaseSeeder',
         ]);
     }
 }
