+++ database/seeders/DatabaseSeeder.php
@@ -8,6 +8,6 @@ class DatabaseSeeder extends Seeder
 {
     public function run(): void
     {
-        $this->call(\Database\Seeds\TestCompany::class);
+        $this->call(\Database\Seeders\TestCompany::class);
     }
 }
