+++ database/seeders/Permissions.php
@@ -1,5 +1,5 @@
 <?php
-namespace Database\Seeds;
+namespace Database\Seeders;
 
 use App\Abstracts\Model;
 use App\Traits\Permissions as Helper;
