+++ modules/Employees/Listeners/Update/Version132.php
@@ -28,7 +28,7 @@ class Version132 extends Listener
     {
         Artisan::call('company:seed', [
             'company' => company_id(),
-            '--class' => 'Modules\Employees\Database\Seeds\Permissions',
+            '--class' => 'Modules\Employees\Database\Seeders\Permissions',
         ]);
 
         $this->detachPermission('employee', 'update-common-items', false);
